+++ packages/node/src/behaviors/operational-credentials/OperationalCredentialsServer.ts
@@ -335,7 +335,7 @@ export class OperationalCredentialsServer extends OperationalCredentialsBehavior
 
         if (timedOp.rootCertSet) {
             // CERTIFICATION BUG WORKAROUND
-            // This should be a ConstraintError but tests require this error
+            // This should be a ConstraintError, but tests require this error
             // See https://github.com/CHIP-Specifications/chip-test-plans/issues/4807
             return {
                 statusCode: OperationalCredentials.NodeOperationalCertStatus.MissingCsr,